+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.8.0]
+### Refactored
+- Remove prop html from react-tooltip instances
+- Sanitize all `data-tip` instances with DOMPurify and add `data-html`. [Reference](https://github.com/wwayne/react-tooltip#security-note)
+
 ## [0.7.4]
 ### Fixed
 - Fix error message style for ValidFieldTextArea